+++ src/cards/support/liu_su.rs
@@ -30,7 +30,7 @@ impl StatusImpl for LiuSu {
         let CommandSource::Switch { dst_char_idx, .. } = e.c.ctx.src else {
             return None;
         };
-        let char_state = &e.c.src_player_state.char_states[dst_char_idx as usize];
+        let char_state = &e.c.src_player_state.char_states[dst_char_idx];
         if char_state.get_energy() < char_state.char_id.get_char_card().max_energy {
             e.add_cmd(Command::AddEnergy(1));
             Some(AppliedEffectResult::ConsumeUsage)
